Added payments

+++ app/Http/Controllers/BookController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Book;
 use App\Models\BookInfo;
 use App\Models\BookItem;
+use App\Models\BookItemInfo;
 use App\services\BookService;
 use Carbon\Carbon;
 use Exception;
@@ -23,6 +24,8 @@ class BookController extends CommonController
     {
         parent::__construct();
         $this->bookService = $service;
+        $this->data['category_name'] = 'Books';
+        $this->data['action_name'] = '';
     }
 
     /**
@@ -35,20 +38,21 @@ class BookController extends CommonController
             return DataTables::of($this->bookService->getBooks())
                 ->addColumn('actions', function ($row){
                     $showUrl = route('books.show',['book' => $row->id]);
+                    $editUrl = route('books.edit',['book' => $row->id]);
                     $canDelete = DB::table('book_items')->where('book_id','=',$row->id)->count() > 0;
                     if($canDelete){
                         return "<a class='btn btn-primary rounded btn-sm text-white font-weight-bold mr-1' href='$showUrl'
                                     data-id='$row->id'>
                                 <i class='fa fa-eye' data-id='$row->id'></i>
                             </a>"
-                            ."<a class='btn-success btn btn-sm rounded text-white  font-weight-bold mr-1 ' onclick='EditBook(event)' data-id='$row->id'>
+                            ."<a class='btn-success btn btn-sm rounded text-white  font-weight-bold mr-1 ' href='$editUrl' data-id='$row->id'>
                                 <i class='fa fa-edit' data-id='$row->id'></i>
                              </a>";
                     }
                     return "<a class='btn btn-primary rounded btn-sm text-white font-weight-bold mr-1' href='$showUrl'>
                                 <i class='fa fa-eye'></i>
                             </a>"
-                        ."<a class='btn-success btn btn-sm rounded text-white  font-weight-bold mr-1 '  onclick='EditBook(event)'>
+                        ."<a class='btn-success btn btn-sm rounded text-white  font-weight-bold mr-1 '  href='$editUrl'>
                                 <i class='fa fa-edit ' data-id='$row->id' data-title='$row->title'></i>
                              </a>"
                         ."<a class='btn btn-danger btn-sm rounded text-white font-weight-bold' onclick='DeleteBook(event)' data-id='$row->id' data-title='$row->title'>
@@ -60,7 +64,12 @@ class BookController extends CommonController
                 ->make(true);
         }
         $this->data['categories'] = DB::table('categories')->select('id','name')->get();
+        $this->data['action_name'] = 'Index';
         return view('books.index')->with('data',$this->data);
+    }
+
+    public function create(Request $request){
+        return view('books.create')->with('data',$this->data);
     }
 
     public function store(Request $request){
@@ -79,9 +88,36 @@ class BookController extends CommonController
         ]);
         $result = $this->bookService->storeBook($data,$request);
         if($result){
-            return response(['message' => 'Book Stored'],201);
+            return redirect(route('books.index'))->with('message','Books Stored');
         }
-        return response(['message' => 'Something went wrong' ],401);
+        return redirect(route('books.index'))->with('error','Something went wrong');
+    }
+
+    public function edit(Book $book){
+        //dd($book);
+        $this->data['book'] = $book;
+        return view('books.edit')->with('data',$this->data);
+    }
+
+    public function update(Request $request){
+        $data = $request->validate([
+            'title' => 'required',
+            'author_id' => 'required',
+            'sale_price' => 'required',
+            'purchase_price' => 'required',
+            'publication_date' => 'required',
+            'age_restricted' => 'required',
+            'num_pages' => 'required',
+            'isbn' => 'required',
+            'cover' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
+            'short_description' => 'required',
+            'categories' => 'required|array'
+        ]);
+        $result = $this->bookService->updateBook($data,$request);
+        if($result){
+            return redirect(route('books.index'))->with('message','Books Stored');
+        }
+        return redirect(route('books.index'))->with('error','Something went wrong');
     }
 
     public function destroy(Request $request){
@@ -89,7 +125,7 @@ class BookController extends CommonController
             'book_id' => 'required'
         ]);
 
-        $result = $this->bookService->deleteBook($data['member_id']);
+        $result = $this->bookService->deleteBook($data);
         if($result){
             return response(['message' => 'Book deleted'],201);
         }
@@ -97,10 +133,12 @@ class BookController extends CommonController
     }
 
     public function show(Request $request, BookInfo $book){
-        if($request->ajax()){
-
-        }
-        return view()->with('data',$this->data);
+        $this->data['book'] = $book;
+        $this->data['categories'] = DB::table('book_category_info')->where('book_id',$book->id)->select('category')->get();
+        $this->data['statuses'] = DB::table('status')->select('id','name')->get();
+        $this->data['conditions'] = DB::table('condition')->select('id','name')->get();
+        //dd($this->data);
+        return view('books.show')->with('data',$this->data);
     }
 
     public function getBookList(Request $request){
@@ -125,7 +163,8 @@ class BookController extends CommonController
             'book_id' => 'required'
         ]);
         $book = Book::findOrFail($data['book_id']);
-        return response()->json(['book'=>$book],201);
+        $categories = DB::table('book_categories')->where('book_id','=', $book->id)->select('id')->get();
+        return response()->json(['book'=>$book,'categories' => $categories],201);
     }
     /**
      * @param Request $request
@@ -136,21 +175,18 @@ class BookController extends CommonController
         if($request->ajax()){
             return DataTables::of($this->bookService->getBookCopies($request->all()))
                 ->addColumn('actions', function ($row){
-                    $showUrl = '';
-//                    $canDelete = DB::table('book_items')->where('book_id','=',$row->id)->count() > 0;
-//                    if($canDelete){
-//                        return "<a class='btn btn-primary rounded btn-sm text-white font-weight-bold mr-1' href='$showUrl'
-//                                    data-id='$row->id'>
-//                                <i class='fa fa-eye' data-id='$row->id'></i>
-//                            </a>"
-//                            ."<a class='btn-success btn btn-sm rounded text-white  font-weight-bold mr-1 ' onclick='EditBook(event)' data-id='$row->id'>
-//                                <i class='fa fa-edit' data-id='$row->id'></i>
-//                             </a>";
-//                    }
-                    return "<a class='btn btn-primary rounded btn-sm text-white font-weight-bold mr-1' href='$showUrl'>
-                                <i class='fa fa-eye'></i>
+                    $showUrl = route('books.copies.show',['bookCopyId' => $row->id]);
+                    $canDelete = $row->status_id == 2;
+                    if($canDelete){
+                        return "<a class='btn btn-primary rounded btn-sm text-white font-weight-bold mr-1' onclick='viewDetails(event)'
+                                    data-id='$row->id'>
+                                <i class='fa fa-eye' data-id='$row->id'></i>
+                            </a>";
+                    }
+                        return "<a class='btn btn-primary rounded btn-sm text-white font-weight-bold mr-1' onclick='viewDetails(event)' data-id='$row->id'>
+                                <i class='fa fa-eye' data-id='$row->id'></i>
                             </a>"
-                        ."<a class='btn-success btn btn-sm rounded text-white  font-weight-bold mr-1 ' data-id='$row->id'   onclick='EditBookCopy(event)'>
+                        ."<a class='btn-success btn btn-sm rounded text-white  font-weight-bold mr-1 ' data-id='$row->id' data-title='$row->title'  onclick='EditBookCopy(event)'>
                                 <i class='fa fa-edit ' data-id='$row->id' data-title='$row->title'></i>
                              </a>"
                         ."<a class='btn btn-danger btn-sm rounded text-white font-weight-bold' onclick='DeleteBookCopy(event)' data-title='$row->title' data-id='$row->id' data-uid='$row->uid'>
@@ -158,12 +194,15 @@ class BookController extends CommonController
                              </a>";
 
                 })
-                ->rawColumns(['actions'])
+                ->addColumn('book_status',function($row){
+                    return $this->getItemStatusColumn($row->status_id,$row->status);
+                })
+                ->rawColumns(['actions','book_status'])
                 ->make(true);
         }
-        $this->data['statuses'] = DB::table('status')->select('id','name')->get();
+        $this->data['statuses'] = DB::table('status')->whereIn('id',[1,2])->select('id','name')->get();
         $this->data['conditions'] = DB::table('condition')->select('id','name')->get();
-        return view('books.items')->with('data',$this->data);
+        return view('books.copy.index')->with('data',$this->data);
     }
 
     public function storeCopy(Request $request){
@@ -179,6 +218,15 @@ class BookController extends CommonController
         }
         return response(['message' => 'Something went wrong' ],401);
     }
+
+    public function showCopy(Request $request, $bookCopyId){
+        $book = DB::table('book_item_info')->where('id','=',$bookCopyId)
+                    ->select('*')->first();
+        $this->data['book_copy'] = $book;
+
+        return view('books.copy.show')->with('data', $this->data);
+    }
+
 
     public function updateCopy(Request $request){
         $data = $request->validate([
@@ -209,7 +257,44 @@ class BookController extends CommonController
         $data = $request->validate([
             'copy_id' => 'required'
         ]);
-        $book = BookItem::findOrFail($data['copy_id']);
+        $book = BookItemInfo::findOrFail($data['copy_id']);
         return response()->json(['book'=>$book],201);
     }
+
+    public function getBookItemList(Request $request){
+        $term = $request['term'];
+        $page = $request['page'];
+        $status = $request['status'] ?? null;
+        $take = 10;
+        $offSet = ($page - 1) * $take;
+
+        $results = DB::table('book_item_info')
+                        ->where('title','like',"%$term%")
+                        ->orWhere('uid','like',"%$term%")
+                        ->select('id',DB::raw( "CONCAT('(', uid,')', ' ', title) as text"));
+        if ($status){
+            $request = $results->where('status_id','=',$status);
+        }
+        $count = $results->count();
+        if(isset($page)){
+            $results = $results->offset($offSet)->take($take);
+        }
+        return response()->json([
+            'results' => $results->get(),
+            'total_item' =>$count
+        ]);
+    }
+
+    /**
+     * @param Request $request
+     * @return mixed
+     * @throws Exception
+     */
+    public function getBookCopyLoansList(Request $request){
+        $id = $request['copy_id'];
+        $loans = DB::table('loan_info')->where('book_item_id','=', $id)
+                ->select('id','member','status','loan_date', 'expected_date')->get();
+        return DataTables::of($loans)->make(true);
+    }
+
 }
